backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
openai
python-decouple